# server/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: API for The Modern Groom, providing a platform to manage users, events,
    and orders for a modern event planning experience.
  title: The Modern Groom API
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /events:
    post:
      operationId: create_event
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
        required: true
        x-body-name: event
      responses:
        "201":
          description: Event created
      summary: Create event
      tags:
      - Events
      - Events Create
      x-openapi-router-controller: openapi_server.controllers.events
    put:
      operationId: update_event
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEvent'
        required: true
        x-body-name: event
      responses:
        "201":
          description: Event Updated Success
      summary: Upadte event
      tags:
      - Events
      - Event Update
      x-openapi-router-controller: openapi_server.controllers.events
  /events/{username}:
    get:
      operationId: list_events
      parameters:
      - explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Successful response
      summary: Get event by email
      tags:
      - Events
      - Event Get
      x-openapi-router-controller: openapi_server.controllers.events
...
  /users:
    get:
      operationId: list_users
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
          description: An array of users
      summary: Lists all users
      tags:
      - Users
      x-openapi-router-controller: openapi_server.controllers.users
    post:
      operationId: create_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
        x-body-name: user_data
      responses:
        "201":
          description: User created
      summary: Create user
      tags:
      - Users
      x-openapi-router-controller: openapi_server.controllers.users
    put:
      operationId: update_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
        x-body-name: user_data
      responses:
        "200":
          description: User updated
        "404":
          description: User not found
      summary: Update a user by ID
      tags:
      - Users
      x-openapi-router-controller: openapi_server.controllers.users
    # delete:
    #   operationId: delete_user
    #   parameters:
    #   - name: email
    #     in: query
    #     description: Unique identifier of the user to delete
    #     required: true
    #     schema:
    #       type: string
    #       format: email
    #   responses:
    #     "204":
    #       description: User deleted successfully
    #     "404":
    #       description: User not found
    #   summary: Delete order by ID
    #   tags:
    #   - Users
    #   - Users Delete
    #   x-openapi-router-controller: openapi_server.controllers.users
  /users/{email}:
    get:
      operationId: get_user_by_id
      parameters:
      - description: Unique ID of the user to retrieve
        explode: false
        in: path
        name: email
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Details of the user
        "404":
          description: User not found
      summary: Get a single user by ID
      tags:
      - Users
      x-openapi-router-controller: openapi_server.controllers.users
...
components:
  schemas:
    User:
      properties:
        id:
          format: uuid
          readOnly: true
          title: id
          type: string
        first_name:
          title: first_name
          type: string
        last_name:
          title: last_name
          type: string
        email:
          title: email
          type: string
        shopify_id:
          title: shopify_id
          type: string
        account_status:
          title: account_status
          type: boolean
        role:
          title: role
          type: string
      required:
        - email
      title: User
      type: object
    Event:
      properties:
        id:
          format: uuid
          readOnly: true
          title: id
          type: string
        event_name:
          format: string
          title: event_name
          type: string
        event_date:
          format: date-time
          title: event_date
          type: string
        user_id:
          format: uuid
          readOnly: true
          title: user_id
          type: string
      required:
        - event_name
        - event_date
      title: Event
      type: object
    UpdateEvent:
      properties:
        id:
          format: uuid
          readOnly: true
          title: id
          type: string
        user_id:
          format: uuid
          readOnly: true
          title: user_id
          type: string
        event_date:
          format: date-time
          title: event_date
          type: string
      required:
        - user_id
        - event_date
      title: Event
      type: object
...